+++ Stitch/App/StitchApp.swift
@@ -44,8 +44,8 @@ struct StitchApp: App {
 #if DEV_DEBUG
     var body: some Scene {
         WindowGroup {
-            VarBodyParserDemoView()
-            // ASTExplorerView()
+            // VarBodyParserDemoView()
+            ASTExplorerView()
         }
     }
 #else
